Refactor aggregate to split out each aggregate function into a separate file
